refactor(recipe): drop CachedStore/RegistryCountForTesting helpers
Deletes the two count-based test inspection helpers from pkg/recipe's
exported surface. They were convention-only test hooks (the _ForTesting
suffix is not enforced) and the only call site
(TestClient_NoCacheGrowthAcrossManyCloseCycles) was a baseline-diff
assertion that conflated this test's cache growth with sibling tests'.

The moved test now lives in the internal aicr test package where it can
read Client.dp directly. It captures each iteration's DataProvider
before Close and asserts Contains(dp) == false per iteration — a
strictly stronger, sibling-test-independent signal than the previous
cumulative count delta.

The surviving Contains*ForTesting helpers remain exported (per-provider
scope, no global cache visibility) and are documented for the same
narrow purpose.

Refs #1079
